v2.5.15: changes to entry editor to default to yes and some other minor tweaks to entry editor and series stats

+++ src/models/book.py
@@ -15,7 +15,7 @@ class Book(Base):
     page_count = Column(Integer)
     date_published = Column(Date)
     series = Column(VARCHAR)
-    series_number = Column(Integer)
+    series_number = Column(Float)
     genre = Column(VARCHAR)
     cover = Column(Boolean, nullable=False, default=False)
     isbn_id = Column(Integer)
